+++ src/main/java/ua/com/dudaryk/repository/interfaces/AbstractDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface AbstractDAO<T> {
 
-    T findById(Long id);
+    T findById(int id);
 
     T save(T entity);
